+++ pokemon-game/src/components/MenuHeader/index.js
@@ -1,0 +1,22 @@
+import { useState } from 'react';
+import Menu from "../Menu";
+import NavBar from "../NavBar";
+
+const MenuHeader = () => {
+    const [isOpenMenu, setOpenMenu] = useState(false);
+
+    const handleVisebleMenu = (isOpenMenu) => {
+        setOpenMenu(!isOpenMenu);
+        console.log("MenuHeader", isOpenMenu);
+    };
+
+    return (
+        <>
+            <Menu isActive={isOpenMenu} />
+            <NavBar onVisibleMenu={handleVisebleMenu}
+            />
+        </>
+    )
+}
+
+export default MenuHeader;
