+++ pokemon-game/src/components/PokemonCard/index.js
@@ -1,8 +1,9 @@
 import { useState } from 'react';
-
-import s from './style.module.css';
+import className from 'classnames';
 
 import cardBackSide from './assets/card-back-side.jpg';
+
+import s from './style.module.css';
 
 const PokemonCard = ({name, img, id, type, values}) => {
     const [isActive, setActive] = useState(false);
@@ -13,15 +14,15 @@ const PokemonCard = ({name, img, id, type, values}) => {
 
     return (
         <div className={s.root} onClick={handleClick}>
-            <div className={`${s.pokemonCard} ${isActive ? s.active : ''}`}>
+            <div className={className(s.pokemonCard, {[s.active] : isActive})}>
                 <div className={s.cardFront}>
-                    <div className={`${s.wrap} ${s.front}`}>
-                        <div className={`${s.pokemon} ${s[type]}`}>
+                    <div className={className(s.wrap, s.front)}>
+                        <div className={className(s.pokemon, s[type])}>
                             <div className={s.values}>
-                                <div className={`${s.count} ${s.top}`}>{values.top}</div>
-                                <div className={`${s.count} ${s.right}`}>{values.right}</div>
-                                <div className={`${s.count} ${s.bottom}`}>{values.bottom}</div>
-                                <div className={`${s.count} ${s.left}`}>{values.left}</div>
+                                <div className={className(s.count, s.top)}>{values.top}</div>
+                                <div className={className(s.count, s.right)}>{values.right}</div>
+                                <div className={className(s.count, s.bottom)}>{values.bottom}</div>
+                                <div className={className(s.count, s.left)}>{values.left}</div>
                             </div>
                             <div className={s.imgContainer}>
                                 <img src={img} alt={name} />
